moved the rubber tree classes to it's own class and added rubber tree leaves
